refactor: move files in directory
